Updated consumer modules structuring
